requirements.txt
asttokens==3.0.0
beautifulsoup4==4.13.5
bertopic==0.17.3
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cssselect==1.3.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dnspython==2.8.0
exceptiongroup==1.3.0
executing==2.2.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
fsspec==2025.9.0
hdbscan==0.8.40
huggingface-hub==0.34.4
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
langdetect==1.0.9
llvmlite==0.44.0
lxml==6.0.1
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
pymongo==4.15.0
pynndescent==0.5.13
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
textblob==0.19.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wordcloud==1.9.4
